Find basin/plateau in 1D array
Find basin/ plateau in 1D array is present. A plateau is defined as a region of increasing
values, followed by at least 2 constant value and a region of decreasing values. A basin is
defined as a region of decreasing values, followed by at least 2 constant value and a region
of increasing values. A list may have a plateau or basin or none.