#days #hours calculation
Calculate the total cost for a pay-as-you-go service availed based on the number of days and
additional hours. The timestamp is taken in ‘dd-mm-yyyy hh:mm:ss’ format. There is a
staring timestamp and an ending timestamp. Calculate the number of days and hours
elapsed between the two timestamps. The chargeable amount for the service needs to be
calculated. The following is the tariff details: 
1 day : Rs. 23
1 hour : Rs. 1